+++ driver/memory/memorykv/store_test.go
@@ -1,0 +1,17 @@
+package memorykv_test
+
+import (
+	"testing"
+
+	. "github.com/dogmatiq/persistencekit/driver/memory/memorykv"
+	"github.com/dogmatiq/persistencekit/kv"
+)
+
+func TestStore(t *testing.T) {
+	kv.RunTests(
+		t,
+		func(t *testing.T) kv.Store {
+			return &Store{}
+		},
+	)
+}