+++ driver/sql/postgres/pgjournal/doc.go
@@ -1,3 +1,3 @@
 // Package pgjournal provides an implementation of [journal.Store] that persists
-// records in a PostgreSQL database.
+// to a PostgreSQL database.
 package pgjournal